+++ core/webdav/methods.go
@@ -1,0 +1,11 @@
+package webdav
+
+const (
+	PROPFIND  = "PROPFIND"
+	MKCOL     = "MKCOL"
+	MOVE      = "MOVE"
+	COPY      = "COPY"
+	LOCK      = "LOCK"
+	UNLOCK    = "UNLOCK"
+	PROPPATCH = "PROPPATCH"
+)